requirements.txt
requests>=2.31.0
python-whois>=0.8.0
tldextract>=3.4.4
dnspython>=2.4.2
